tempodisco: Temporal Discounting Models

Tools for working with temporal discounting data, designed for behavioural researchers to simplify data cleaning/scoring and model fitting. The package implements widely used methods such as computing indifference points from adjusting amount task (Frye et al., 2016, doi:10.3791/53584 ), testing for non-systematic discounting per the criteria of Johnson & Bickel (2008, doi:10.1037/1064-1297.16.3.264 ), scoring questionnaires according to the methods of Kirby et al. (1999, doi:10.1037//0096-3445.128.1.78 ) and Wileyto et al (2004, doi:10.3758/BF03195548 ), Bayesian model selection using a range of discount functions (Franck et al., 2015, doi:10.1002/jeab.128 ), drift diffusion models of discounting (Peters & D'Esposito, 2020, doi:10.1371/journal.pcbi.1007615 ), and model-agnostic measures of discounting such as area under the curve (Myerson et al., 2001, doi:10.1901/jeab.2001.76-235 ) and ED50 (Yoon & Higgins, 2008, doi:10.1016/j.drugalcdep.2007.12.011 ).

See also

Author

Maintainer: Isaac Kinley isaac.kinley@gmail.com